requirements.txt
ConcurrentLogHandler==0.9.1
pyyaml==5.1.2
pymongo==3.9.0
python_dateutil==2.8.0
pytz==2017.3
six==1.12.0
numpy==1.17.0
pandas==0.25.0
mpmath==1.1.0
sympy==1.4
click==7.0
MarkupSafe==1.1.1
Jinja2==2.10.1
Werkzeug==0.15.5
itsdangerous==1.1.0
flask==1.1.1
scipy==1.3.1
scikit-learn==0.20.2
flask_cors==3.0.8
pytorch==1.0.1.post2
urllib3==1.25.3
idna==2.8
certifi==2019.6.16
chardet==3.0.4
requests==2.22.0